@@ -11,15 +11,15 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "17.3.8",
+    "@angular/animations": "18.0.0",
     "@angular/cdk": "^17.3.8",
-    "@angular/common": "17.3.8",
-    "@angular/compiler": "17.3.8",
-    "@angular/core": "17.3.8",
-    "@angular/forms": "17.3.8",
-    "@angular/platform-browser": "17.3.8",
-    "@angular/platform-browser-dynamic": "17.3.8",
-    "@angular/router": "17.3.8",
+    "@angular/common": "18.0.0",
+    "@angular/compiler": "18.0.0",
+    "@angular/core": "18.0.0",
+    "@angular/forms": "18.0.0",
+    "@angular/platform-browser": "18.0.0",
+    "@angular/platform-browser-dynamic": "18.0.0",
+    "@angular/router": "18.0.0",
     "primeflex": "^3.2.1",
     "primeicons": "^6.0.1",
     "primeng": "^17.16.1",
@@ -30,8 +30,8 @@
   "devDependencies": {
     "@angular-devkit/build-angular": "^17.3.7",
     "@angular/cli": "^17.3.7",
-    "@angular/compiler-cli": "17.3.8",
-    "@angular/language-service": "17.3.8",
+    "@angular/compiler-cli": "18.0.0",
+    "@angular/language-service": "18.0.0",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.8",
     "@types/node": "^12.12.38",
@@ -46,6 +46,6 @@
     "protractor": "~7.0.0",
     "ts-node": "~8.8.2",
     "tslint": "~6.1.0",
-    "typescript": "^5.2.2"
+    "typescript": "^5.4.5"
   }
 }
